feat: bump rgb version
