+++ Cargo.toml
@@ -27,7 +27,6 @@ server = ["tokio/full", "tower-http/cors"]
 
 [dependencies]
 anyhow = "1.0.71"
-aluvm = "0.10.2"
 amplify = "4.0.0"
 argon2 = "0.5.0"
 base64 = { package = "base64-compat", version = "1.0.0" }
@@ -39,9 +38,9 @@ bitcoin = { version = "0.29.2", features = ["base64"] }
 bitcoin_hashes = "0.12.0"
 bitcoin_scripts = "0.10.0-alpha.2"
 bitcoin_blockchain = "0.10.0-alpha.2"
-bp-core = "0.10.3"
-commit_verify = "0.10.2"
-bp-seals = "0.10.3"
+bp-core = "0.10.4"
+commit_verify = "0.10.3"
+bp-seals = "0.10.4"
 indexmap = "1.9.3"
 carbonado = "0.3.0-rc.13"
 console_error_panic_hook = "0.1.7"
@@ -74,18 +73,22 @@ psbt = { version = "0.10.0-alpha.2", features = [
 ] }
 regex = "1.7.0"
 reqwest = { version = "0.11.16", features = ["json"] }
-rgb-contracts = { version = "0.10.0", default-features = false, features = [
+rgb-contracts = { version = "0.10.0-beta.2", default-features = false, features = [
     "log",
 ] }
-rgb-std = { version = "0.10.2" }
 rgb-wallet = { version = "0.10.2" }
+rgb-std = { version = "0.10.2" }
+
+# rgb_wallet_bp = { version = "0.10.2-backport", package = "rgb-wallet", path = "../rgb-wallet" }
+# rgb_std_bp = { version = "0.10.2-backport", package = "rgb-std", path = "../rgb-wallet/std" }
+
 rgb-schemata = { version = "0.10.0-beta.1" }
 serde = "1.0.152"
 serde_json = "1.0.91"
 postcard = { version = "1.0.4", features = ["alloc"] }
 serde-encrypt = "0.7.0"
-strict_types = "1.2.0"
-strict_encoding = "2.2.0"
+strict_types = "1.3.0"
+strict_encoding = "2.3.0"
 tokio = { version = "1.28.2", features = ["macros", "sync"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
@@ -94,7 +97,7 @@ bdk = { version = "0.28.0", features = [
     "async-interface",
 ], default-features = false }
 gloo-console = "0.2.3"
-gloo-net = "0.2.6"
+gloo-net = "0.3.0"
 js-sys = "0.3.63"
 serde-wasm-bindgen = "0.5.0"
 wasm-bindgen = { version = "0.2.86", features = ["serde-serialize"] }
@@ -124,14 +127,13 @@ wasm-bindgen-test = "0.3.36"
 
 [patch.crates-io]
 # TODO: Remove this after merge and release fixies
-rgb-contracts = { git = "https://github.com/crisdut/rgb", branch = "exp/fixies" }
 bitcoin_scripts = { git = "https://github.com/crisdut/bp-foundation", branch = "release/0.10.0-alpha.2" }
 bitcoin_blockchain = { git = "https://github.com/crisdut/bp-foundation", branch = "release/0.10.0-alpha.2" }
-strict_types = { git = "https://github.com/crisdut/strict-types/", branch = "fix/typelib-id" }
 # TODO: Remove this after propose a universal tweak in descriptor-wallet
 descriptor-wallet = { git = "https://github.com/crisdut/descriptor-wallet", branch = "exp/tapret-transfers" }
 psbt = { git = "https://github.com/crisdut/descriptor-wallet", branch = "exp/tapret-transfers" }
 # TODO: Remove this after merge and release fixies
-rgb-schemata = { git = "https://github.com/crisdut/rgbsc", branch = "exp/full-uda" }
-rgb-wallet = { git = "https://github.com/crisdut/rgb-wallet", branch = "exp/full-uda" }
-rgb-std = { git = "https://github.com/crisdut/rgb-wallet", branch = "exp/full-uda" }
+rgb-contracts = { git = "https://github.com/crisdut/rgb", branch = "release/bmc-v0.6" }
+rgb-schemata = { git = "https://github.com/crisdut/rgbsc", branch = "release/bmc-v0.6" }
+rgb-wallet = { git = "https://github.com/crisdut/rgb-wallet", branch = "release/bmc-v0.6" }
+rgb-std = { git = "https://github.com/crisdut/rgb-wallet", branch = "release/bmc-v0.6" }
