+++ tests/rgb.rs
@@ -13,12 +13,13 @@ mod rgb {
         // mod collectibles;
         mod collectibles;
         mod fungibles;
-        mod import;
         mod issue;
+        mod states;
         mod stress;
         mod udas;
         pub mod utils;
         mod watcher;
+        mod import;
     }
 
     mod web {
