+++ tests/rgb/integration/import.rs
@@ -1,250 +1,48 @@
 #![cfg(not(target_arch = "wasm32"))]
+use crate::rgb::integration::utils::ISSUER_MNEMONIC;
 use bitmask_core::{
-    bitcoin::{save_mnemonic, sign_psbt_file},
-    rgb::{accept_transfer, create_watcher, get_contract},
-    structs::{AcceptRequest, DecryptedWalletData, SecretString, SignPsbtRequest, WatcherRequest},
-};
-
-use crate::rgb::integration::utils::{
-    create_new_invoice, create_new_psbt, create_new_transfer, get_uda_data, import_new_contract,
-    issuer_issue_contract, ISSUER_MNEMONIC, OWNER_MNEMONIC,
+    bitcoin::save_mnemonic,
+    rgb::import,
+    structs::{AssetType, ImportRequest, SecretString},
 };
 
 #[tokio::test]
 async fn allow_import_fungible_contract() -> anyhow::Result<()> {
-    let issuer_resp = issuer_issue_contract("RGB20", 5, false, true, None).await;
-    assert!(issuer_resp.is_ok());
+    let issuer_keys = save_mnemonic(
+        &SecretString(ISSUER_MNEMONIC.to_string()),
+        &SecretString("".to_string()),
+    )
+    .await?;
 
-    let import_resp = import_new_contract(issuer_resp?).await;
+    let sk = &issuer_keys.private.nostr_prv;
+    let contract_import = ImportRequest {
+        import: AssetType::RGB20,
+        data: FUNGIBLE_CONTRACT.to_string(),
+    };
+
+    let import_resp = import(sk, contract_import).await;
     assert!(import_resp.is_ok());
     Ok(())
 }
 
 #[tokio::test]
 async fn allow_import_uda_contract() -> anyhow::Result<()> {
-    let single = Some(get_uda_data());
-    let issuer_resp = issuer_issue_contract("RGB21", 1, false, true, single).await;
-    assert!(issuer_resp.is_ok());
+    let issuer_keys = save_mnemonic(
+        &SecretString(ISSUER_MNEMONIC.to_string()),
+        &SecretString("".to_string()),
+    )
+    .await?;
 
-    let import_resp = import_new_contract(issuer_resp?).await;
+    let sk = &issuer_keys.private.nostr_prv;
+    let contract_import = ImportRequest {
+        import: AssetType::RGB21,
+        data: UDA_CONTRACT.to_string(),
+    };
+
+    let import_resp = import(sk, contract_import).await;
     assert!(import_resp.is_ok());
     Ok(())
 }
 
-// TODO: Review after support multi-token transfer
-// async fn _allow_import_collectible_contract() -> anyhow::Result<()> {
-//     let collectibles = Some(get_collectible_data());
-//     let issuer_resp = issuer_issue_contract("RGB21", 1, false, true, collectibles).await;
-//     assert!(issuer_resp.is_ok());
-
-//     let import_resp = import_new_contract(issuer_resp?).await;
-//     assert!(import_resp.is_ok());
-//     Ok(())
-// }
-
-#[tokio::test]
-async fn allow_get_fungible_contract_state_by_accept_cosign() -> anyhow::Result<()> {
-    // 1. Issue and Generate Trasnfer (Issuer side)
-    let issuer_keys: DecryptedWalletData = save_mnemonic(
-        &SecretString(ISSUER_MNEMONIC.to_string()),
-        &SecretString("".to_string()),
-    )
-    .await?;
-    let owner_keys = save_mnemonic(
-        &SecretString(OWNER_MNEMONIC.to_string()),
-        &SecretString("".to_string()),
-    )
-    .await?;
-    let issuer_resp = issuer_issue_contract("RGB20", 5, false, true, None).await?;
-    let owner_resp = create_new_invoice(issuer_resp.clone(), None).await?;
-    let psbt_resp = create_new_psbt(issuer_keys.clone(), issuer_resp.clone()).await?;
-    let transfer_resp = &create_new_transfer(issuer_keys.clone(), owner_resp, psbt_resp).await?;
-
-    // 2. Sign and Publish TX (Issuer side)
-    let issuer_sk = issuer_keys.private.nostr_prv.to_string();
-    let owner_sk = owner_keys.private.nostr_prv.to_string();
-    let request = SignPsbtRequest {
-        psbt: transfer_resp.psbt.clone(),
-        descriptor: SecretString(issuer_keys.private.rgb_assets_descriptor_xprv.clone()),
-    };
-    let resp = sign_psbt_file(request).await;
-    assert!(resp.is_ok());
-
-    // 3. Accept Consig (Issuer Side)
-    let request = AcceptRequest {
-        consignment: transfer_resp.clone().consig,
-        force: false,
-    };
-    let resp = accept_transfer(&issuer_sk, request).await;
-    assert!(resp.is_ok());
-    assert!(resp?.valid);
-
-    // 4. Accept Consig (Owner Side)
-    let request = AcceptRequest {
-        consignment: transfer_resp.consig.clone(),
-        force: false,
-    };
-    let resp = accept_transfer(&owner_sk, request).await;
-    assert!(resp.is_ok());
-    assert!(resp?.valid);
-
-    // 5. Retrieve Contract (Issuer Side)
-    let contract_id = &issuer_resp.contract_id;
-    let resp = get_contract(&issuer_sk, contract_id).await;
-    assert!(resp.is_ok());
-    assert_eq!(4, resp?.balance);
-
-    // 6. Create Watcher (Owner Side)
-    let watcher_name = "default";
-    let create_watch_req = WatcherRequest {
-        name: watcher_name.to_string(),
-        xpub: owner_keys.public.watcher_xpub.clone(),
-        force: true,
-    };
-    create_watcher(&owner_sk, create_watch_req).await?;
-
-    // 7. Retrieve Contract (Owner Side)
-    let contract_id = &issuer_resp.contract_id;
-    let resp = get_contract(&owner_sk, contract_id).await;
-    assert!(resp.is_ok());
-    assert_eq!(1, resp?.balance);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn allow_get_uda_contract_state_by_accept_cosign() -> anyhow::Result<()> {
-    // 1. Issue and Generate Trasnfer (Issuer side)
-    let single = Some(get_uda_data());
-    let issuer_keys: DecryptedWalletData = save_mnemonic(
-        &SecretString(ISSUER_MNEMONIC.to_string()),
-        &SecretString("".to_string()),
-    )
-    .await?;
-    let owner_keys = save_mnemonic(
-        &SecretString(OWNER_MNEMONIC.to_string()),
-        &SecretString("".to_string()),
-    )
-    .await?;
-
-    let issuer_resp = issuer_issue_contract("RGB21", 1, false, true, single).await?;
-    let owner_resp = create_new_invoice(issuer_resp.clone(), None).await?;
-    let psbt_resp = create_new_psbt(issuer_keys.clone(), issuer_resp.clone()).await?;
-    let transfer_resp = &create_new_transfer(issuer_keys.clone(), owner_resp, psbt_resp).await?;
-
-    // 2. Sign and Publish TX (Issuer side)
-    let issuer_sk = issuer_keys.private.nostr_prv.to_string();
-    let owner_sk = owner_keys.private.nostr_prv.to_string();
-    let request = SignPsbtRequest {
-        psbt: transfer_resp.psbt.clone(),
-        descriptor: SecretString(issuer_keys.private.rgb_udas_descriptor_xprv.clone()),
-    };
-    let resp = sign_psbt_file(request).await;
-    assert!(resp.is_ok());
-
-    // 3. Accept Consig (Issuer Side)
-    let request = AcceptRequest {
-        consignment: transfer_resp.consig.clone(),
-        force: false,
-    };
-    let resp = accept_transfer(&issuer_sk, request).await;
-    assert!(resp.is_ok());
-    assert!(resp?.valid);
-
-    // 4. Accept Consig (Owner Side)
-    let request = AcceptRequest {
-        consignment: transfer_resp.consig.clone(),
-        force: false,
-    };
-    let resp = accept_transfer(&owner_sk, request).await;
-    assert!(resp.is_ok());
-    assert!(resp?.valid);
-
-    // 5. Retrieve Contract (Issuer Side)
-    let contract_id = &issuer_resp.contract_id;
-    let resp = get_contract(&issuer_sk, contract_id).await;
-    assert!(resp.is_ok());
-    assert_eq!(0, resp?.balance);
-
-    // 6. Create Watcher (Owner Side)
-    let watcher_name = "default";
-    let create_watch_req = WatcherRequest {
-        name: watcher_name.to_string(),
-        xpub: owner_keys.public.watcher_xpub.clone(),
-        force: true,
-    };
-    create_watcher(&owner_sk, create_watch_req).await?;
-
-    // 7. Retrieve Contract (Owner Side)
-    let contract_id = &issuer_resp.contract_id;
-    let resp = get_contract(&owner_sk, contract_id).await;
-    assert!(resp.is_ok());
-    assert_eq!(1, resp?.balance);
-
-    Ok(())
-}
-
-// TODO: Review after support multi-token transfer
-// async fn _allow_get_collectible_contract_state_by_accept_cosign() -> anyhow::Result<()> {
-//     // 1. Issue and Generate Trasnfer (Issuer side)
-//     let collectible = Some(get_collectible_data());
-//     let issuer_keys: DecryptedWalletData = save_mnemonic(ISSUER_MNEMONIC, "").await?;
-//     let owner_keys = save_mnemonic(OWNER_MNEMONIC, "").await?;
-//     let issuer_resp = issuer_issue_contract("RGB21", 1, false, true, collectible).await?;
-//     let owner_resp = create_new_invoice(issuer_resp.clone(), None).await?;
-//     let psbt_resp = create_new_psbt(issuer_keys.clone(), issuer_resp.clone()).await?;
-//     let transfer_resp = create_new_transfer(issuer_keys.clone(), owner_resp, psbt_resp).await?;
-
-//     // 2. Sign and Publish TX (Issuer side)
-//     let issuer_sk = issuer_keys.private.nostr_prv.to_string();
-//     let owner_sk = owner_keys.private.nostr_prv.to_string();
-//     let request = SignPsbtRequest {
-//         psbt: transfer_resp.clone().psbt,
-//         mnemonic: ISSUER_MNEMONIC.to_string(),
-//         seed_password: String::new(),
-//         iface: issuer_resp.iface,
-//     };
-//     let resp = sign_psbt_file(&issuer_sk, request).await;
-//     assert!(resp.is_ok());
-
-//     // 3. Accept Consig (Issuer Side)
-//     let request = AcceptRequest {
-//         consignment: transfer_resp.clone().consig,
-//         force: false,
-//     };
-//     let resp = accept_transfer(&issuer_sk, request).await;
-//     assert!(resp.is_ok());
-//     assert!(resp?.valid);
-
-//     // 4. Accept Consig (Owner Side)
-//     let request = AcceptRequest {
-//         consignment: transfer_resp.consig,
-//         force: false,
-//     };
-//     let resp = accept_transfer(&owner_sk, request).await;
-//     assert!(resp.is_ok());
-//     assert!(resp?.valid);
-
-//     // 5. Retrieve Contract (Issuer Side)
-//     let contract_id = &issuer_resp.contract_id;
-//     let resp = get_contract(&issuer_sk, contract_id).await;
-//     assert!(resp.is_ok());
-//     assert_eq!(0, resp?.balance);
-
-//     // 6. Create Watcher (Owner Side)
-//     let watcher_name = "default";
-//     let create_watch_req = WatcherRequest {
-//         name: watcher_name.to_string(),
-//         xpub: owner_keys.public.watcher_xpub,
-//         force: true,
-//     };
-//     create_watcher(&owner_sk, create_watch_req).await?;
-
-//     // 7. Retrieve Contract (Owner Side)
-//     let contract_id = &issuer_resp.contract_id;
-//     let resp = get_contract(&owner_sk, contract_id).await;
-//     assert!(resp.is_ok());
-//     assert_eq!(1, resp?.balance);
-
-//     Ok(())
-// }
+const FUNGIBLE_CONTRACT: &str = "0100000000030300596e7a3a0ea3659e1030f18b4d65a3b913ee585007fcf75e0bb6cdc169c1c03b0100d00799624db7e426929954932edfc8d267a5529702293cd5abc0564e6f225b33faf10100d10718cb946f1293cf180e9d78dcc65bc59b472ffffeadfbf58db198cc8328f64b01010001d007010800d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c03030001000100d00701000100d1070100010001d0070100ffff000001d007d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0001d0070100ffff01d0070100ffff003c000001196d8529c8f6da444d154e8d9c8cbd9e0f8d89025a3bb09121c63d52b01e870501a5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed707fc7e770beea38fc53848e2890f75b66935f005a2e954bd989f4c099392b0a5081cabbfc3d826c0bfd1e9770a889efacc8b6716ad014a3eec10b65915302290420000000000000000ffffffffffffffff0b6c500d94264462fb71c0e87478037695b4646158b9ec519dd6d47e74dc54f3035f05737061636520046578636c210671756f7465732204686173682306646f6c6c6172240770657263656e742509616d70657273616e64260a61706f7374726f7068652708627261636b65744c2808627261636b6574522908617374657269736b2a04706c75732b05636f6d6d612c056d696e75732d03646f742e05736c6173682f047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e653905636f6c6f6e3a0973656d69436f6c6f6e3b046c6573733c05657175616c3d07677265617465723e087175657374696f6e3f02617440014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a0a7371427261636b65744c5b096261636b536c6173685c0a7371427261636b6574525d0563617265745e066c6f646173685f086261636b7469636b60016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7a0963427261636b65744c7b04706970657c0963427261636b6574527d0574696c64657e0ed21a3b1985f8f55f1affc7261d551f6bab4bf206cdc448fdcb97b7c44809cd0602047574786fe86a40cd5c25f9bde284ff69a51132a7412b2155bd0e5612c1bfe3849560c5060570726f6f66f2ba2480169e7d3a48fd00302d1fdf80b9e3118f01a25ba6752d7dcf81ef083815dfa1833ae7a0256426366659a343d3ca0308da6670cfd663f19878ef9f7465087a7b8f49025cafad1370f90a4b0bb78028c417521384c039b409386768ed91140000000000000000ffffffffffffffff181aa1e3ed74bd51a3ccdb6cb2180bd413f2f891c28a765f16a2a8f3ba2dabb1086015c0e70764a9f48ae71dd448caf4035bffebdb62eff70dc8794286f8910f410100000000000000080000000000000018cb946f1293cf180e9d78dcc65bc59b472ffffeadfbf58db198cc8328f64b010501560d96f7a47924b2c3df040e6463398fd65fd591652c294342bfa5f9391551541cabbfc3d826c0bfd1e9770a889efacc8b6716ad014a3eec10b659153022904200401deddddb38b837060353fb361533ffa352f30397b4c7abc15bbdbd25da8915900501e15eba16722843cc10954a1b33479dbc343c67693fb7ade855e17f656b9f583e21e33e43c5aa5cf2292753a338c167ed3b639c4dd9cebda98f346917bac943750501a5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed72a5baaac5089ff098d150b482cfed8bcd01a91c0d7b45d32216ed576ab71ebdd071cabbfc3d826c0bfd1e9770a889efacc8b6716ad014a3eec10b659153022904220002d2519390702d9a3484b79e289e7bbf2a8cad7fb5d2bc0a5f8471dba6fa3c03c031a014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a30a0d30c6bbfe91e2490b34278c7bf7f66ee6444011888dc4132bf304c6fe7680807fc7e770beea38fc53848e2890f75b66935f005a2e954bd989f4c099392b0a50000000000000000ffffffffffffffff332de353e616da0cb6394a23d10773fae94ff131f4fe216b2d0001bff1c7b2fb06040776657273696f6ea87cc6796467e557d744e5c9623ae00ccb95d91a0d2f03df74c075124fbec79e06696e7075747315dfa1833ae7a0256426366659a343d3ca0308da6670cfd663f19878ef9f7465076f7574707574738ac60209af0af69c70a282f79ba38ddf4ab5a2afbca5da03dc69ffab00e395ca086c6f636b54696d6535da1d1539206ebbdddca4c33d8731eaf29ba7da74df3dc882201c4e16b5b91c339840868f9caa65acd0874d1a06eb2ee2614cdcd537f77615c836cb9a613d93033f04646173682d047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7a35da1d1539206ebbdddca4c33d8731eaf29ba7da74df3dc882201c4e16b5b91c0501a5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed73c8add893836f8c04c073d1f454d9525cf96a7d7557296fa663ddbaaefa843940334014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7a3f2b72b7c4af1a630cb6d3ff088baf351093ee465b9e7d472a610082e449d7e4050163aa2314e8b147c8b284dfb39a9e10d19caad5faea848e3cb9849d9167d6344a45b780258601c526b23b5b4861460a9050e13f35fbbe8305a8001157e4013888040200046e6f6e65d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0104736f6d65caff8faeb38a00a04e3621538f8e61d75a85a465cb0a0e48c3593e7eaa6c5fc75014105f0c4026ca12abae377de26f8e3ca0608dd36c483ed70f47195cd334d4050184de03c07d1d477bb76d40a947d26e1ed87dc44eaf86720dd4dc8a48047373b3560d96f7a47924b2c3df040e6463398fd65fd591652c294342bfa5f93915515408fba958721a3d335406b368c36f5a82790960cce239febcafe189ba9839d5da780000000000000000ffff000000000000596e7a3a0ea3659e1030f18b4d65a3b913ee585007fcf75e0bb6cdc169c1c03b05016632da64bbf5bc3cc9f522452f5377a83804cdcd95006bc467a1394cc1a2b48f5be69aabb827ae7a16a0c65f72ee867e327e63039aa8d51b176a5a1410b2e269050107fc7e770beea38fc53848e2890f75b66935f005a2e954bd989f4c099392b0a55ca149585de534ee91b3e3a030b7efd4cdb79abea9152f101f3759b4c7210e1f00085d03c4178da98e7e3f3af343e3997d74201d11f42732cfbea2b04b8e3ff15f2203130b696e646976697369626c65000464656369010563656e746902056d696c6c690309646563694d696c6c69040a63656e74694d696c6c6905056d6963726f0609646563694d6963726f070a63656e74694d6963726f08046e616e6f0908646563694e616e6f0a0963656e74694e616e6f0b047069636f0c08646563695069636f0d0963656e74695069636f0e0566656d746f0f096465636946656d746f100a63656e746946656d746f11046174746f126015c0e70764a9f48ae71dd448caf4035bffebdb62eff70dc8794286f8910f410324047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a618622d17baef06602dfe775c980e0b36b6ebae8090d8075f7513bec4b93f64d03020566616c73650004747275650163aa2314e8b147c8b284dfb39a9e10d19caad5faea848e3cb9849d9167d6344a08fba958721a3d335406b368c36f5a82790960cce239febcafe189ba9839d5da780100000000000000ff0000000000000065eb913c17e7e03d9bb1c615293b085ade1511a7ba1c6b30160a33722231a10a06020576616c756597f5d70642ae2a3392253ba84e1dcec493e3bf3ed371b1877386354ecc88826b0c7363726970745075626b6579cae95645ac9ae8fb1892420d1c8b46c220aac64ce6dc8258866fa365aea490436632da64bbf5bc3cc9f522452f5377a83804cdcd95006bc467a1394cc1a2b48f0044728ea68a97de7420f62a5969b04b872604bf45d6df38959c7e26e7a72490e4ad033f047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a066c6f646173685f016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7a7a7b8f49025cafad1370f90a4b0bb78028c417521384c039b409386768ed911406040a707265764f7574707574e86a40cd5c25f9bde284ff69a51132a7412b2155bd0e5612c1bfe3849560c50609736967536372697074cbb261c09079bea5f78800d747feb5f8c024e2c04f73f945b468068df143ec940873657175656e636501196d8529c8f6da444d154e8d9c8cbd9e0f8d89025a3bb09121c63d52b01e87077769746e657373ef9b75968c7d8d6905f17c692061d6e93a2abe8248e6a70b0867a2dffe84fb258057ae07f47b923accc5361fcf40d92982e841f757dd59b16d7bdbc8eabe99a50501181aa1e3ed74bd51a3ccdb6cb2180bd413f2f891c28a765f16a2a8f3ba2dabb184de03c07d1d477bb76d40a947d26e1ed87dc44eaf86720dd4dc8a48047373b3080b6c500d94264462fb71c0e87478037695b4646158b9ec519dd6d47e74dc54f3010000000000000028000000000000008ac60209af0af69c70a282f79ba38ddf4ab5a2afbca5da03dc69ffab00e395ca0865eb913c17e7e03d9bb1c615293b085ade1511a7ba1c6b30160a33722231a10a0000000000000000ffffffffffffffff97f5d70642ae2a3392253ba84e1dcec493e3bf3ed371b1877386354ecc88826b05015ca149585de534ee91b3e3a030b7efd4cdb79abea9152f101f3759b4c7210e1f99624db7e426929954932edfc8d267a5529702293cd5abc0564e6f225b33faf10602066e616d696e67de7c753c6bd56774f87245306ca463409710cee01bedca9120fcaeef93bff58609707265636973696f6e5d03c4178da98e7e3f3af343e3997d74201d11f42732cfbea2b04b8e3ff15f229e14cce6b185cca477ba44634bfa2c5871a97e0c32b7aec1df6a94522a9319c70310025f3000025f3101025f3202025f3303025f3404025f3505025f3606025f3707025f3808025f3909035f31300a035f31310b035f31320c035f31330d035f31340e035f31350fa38242f3cbd7d1d0a112da8de1602f56d4f86f0ec4c476c5c068594c4b04655605012a5baaac5089ff098d150b482cfed8bcd01a91c0d7b45d32216ed576ab71ebdda5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed70004a62e384a135184183b9da6b9796b06e88c75c6803139dd391148ea049ae3d7e10310047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e65390374656e6106656c6576656e62067477656c76656308746869727465656e6408666f75727465656e65076669667465656e66a87cc6796467e557d744e5c9623ae00ccb95d91a0d2f03df74c075124fbec79e05016632da64bbf5bc3cc9f522452f5377a83804cdcd95006bc467a1394cc1a2b48fbaa0e3fcab9d55d168bb1e26837153afe89f83a27e0498554d3963f6431fd3f4033e047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7abc2891b1c66ac5f5e61059a32077daa82333a1435bd34608b30b03fc017d9545030a047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539c0ec009baa44e89cdbeb95c6f18117a53618352f070bc59fc1d8e49f07b4298c040200046e6f6e65d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0104736f6d651deddddb38b837060353fb361533ffa352f30397b4c7abc15bbdbd25da891590c3cad9212706689af885a1b06a8d27ed3d3ff250c409f93fee32eb32544866340360035f333220035f333321035f333422035f333523035f333624035f333725035f333826035f333927035f343028035f343129035f34322a035f34332b035f34342c035f34352d035f34362e035f34372f035f343830035f343931035f353032035f353133035f353234035f353335035f353436035f353537035f353638035f353739035f35383a035f35393b035f36303c035f36313d035f36323e035f36333f035f363440035f363541035f363642035f363743035f363844035f363945035f373046035f373147035f373248035f373349035f37344a035f37354b035f37364c035f37374d035f37384e035f37394f035f383050035f383151035f383252035f383353035f383454035f383555035f383656035f383757035f383858035f383959035f39305a035f39315b035f39325c035f39335d035f39345e035f39355f035f393660035f393761035f393862035f393963045f31303064045f31303165045f31303266045f31303367045f31303468045f31303569045f3130366a045f3130376b045f3130386c045f3130396d045f3131306e045f3131316f045f31313270045f31313371045f31313472045f31313573045f31313674045f31313775045f31313876045f31313977045f31323078045f31323179045f3132327a045f3132337b045f3132347c045f3132357d045f3132367e045f3132377fcae95645ac9ae8fb1892420d1c8b46c220aac64ce6dc8258866fa365aea4904305015be69aabb827ae7a16a0c65f72ee867e327e63039aa8d51b176a5a1410b2e269caff8faeb38a00a04e3621538f8e61d75a85a465cb0a0e48c3593e7eaa6c5fc705013f2b72b7c4af1a630cb6d3ff088baf351093ee465b9e7d472a610082e449d7e4cbb261c09079bea5f78800d747feb5f8c024e2c04f73f945b468068df143ec9405015be69aabb827ae7a16a0c65f72ee867e327e63039aa8d51b176a5a1410b2e269d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0000da57a3e7de97d8c4dcaafbd143a882a7587e35c3fa805625098c97e71629a5b408c3cad9212706689af885a1b06a8d27ed3d3ff250c409f93fee32eb325448663401000000000000004000000000000000de7c753c6bd56774f87245306ca463409710cee01bedca9120fcaeef93bff5860603067469636b65728057ae07f47b923accc5361fcf40d92982e841f757dd59b16d7bdbc8eabe99a5046e616d655014105f0c4026ca12abae377de26f8e3ca0608dd36c483ed70f47195cd334d40764657461696c7345b780258601c526b23b5b4861460a9050e13f35fbbe8305a8001157e4013888e15eba16722843cc10954a1b33479dbc343c67693fb7ade855e17f656b9f583e0501da57a3e7de97d8c4dcaafbd143a882a7587e35c3fa805625098c97e71629a5b4e86a40cd5c25f9bde284ff69a51132a7412b2155bd0e5612c1bfe3849560c50606020474786964a38242f3cbd7d1d0a112da8de1602f56d4f86f0ec4c476c5c068594c4b04655604766f757421e33e43c5aa5cf2292753a338c167ed3b639c4dd9cebda98f346917bac94375eedec3ebaa08fb010f2f96e17a0e039ad87fe60771a7301f8c5e88995d05676b0310047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e65390374656e4106656c6576656e42067477656c76654308746869727465656e4408666f75727465656e45076669667465656e46ef9b75968c7d8d6905f17c692061d6e93a2abe8248e6a70b0867a2dffe84fb25050130a0d30c6bbfe91e2490b34278c7bf7f66ee6444011888dc4132bf304c6fe768f2ba2480169e7d3a48fd00302d1fdf80b9e3118f01a25ba6752d7dcf81ef0838081cabbfc3d826c0bfd1e9770a889efacc8b6716ad014a3eec10b65915302290420000000000000000ffff000000000000f9170804ddae0479f8d5af74ab3bd202e6ea4172d9a9b93707151adb7fc40ca1031a016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7afb6b2ec4700e9bf1cd0f24bb9ced0dd78dbb4e1dcaf5c68fa6935a7a4fdc694206030474797065e15eba16722843cc10954a1b33479dbc343c67693fb7ade855e17f656b9f583e0773756274797065c0ec009baa44e89cdbeb95c6f18117a53618352f070bc59fc1d8e49f07b4298c0763686172736574c0ec009baa44e89cdbeb95c6f18117a53618352f070bc59fc1d8e49f07b4298cfba958721a3d335406b368c36f5a82790960cce239febcafe189ba9839d5da78010001bab043436d5163f8872caa19c3c863c0878364fe504a1c6dcfb0f2d2c62dcfdb0c00035247420300d0d007000000010004d007bab043436d5163f8872caa19c3c863c0878364fe504a1c6dcfb0f2d2c62dcfdb0000019abcb19a89c3e955b796ff9044c1078e81ab142c309d8524c774094a1835838f00055247423230060c6275726e6564537570706c7901888c5865633af13b95b7cd1a8d8af2dac1dc140b977251d9d4daf3c7511c8e840001076372656174656401596e7a3a0ea3659e1030f18b4d65a3b913ee585007fcf75e0bb6cdc169c1c03b01000c697373756564537570706c7901888c5865633af13b95b7cd1a8d8af2dac1dc140b977251d9d4daf3c7511c8e8401010e7265706c61636564537570706c7901888c5865633af13b95b7cd1a8d8af2dac1dc140b977251d9d4daf3c7511c8e84000104737065630199624db7e426929954932edfc8d267a5529702293cd5abc0564e6f225b33faf10100057465726d730118cb946f1293cf180e9d78dcc65bc59b472ffffeadfbf58db198cc8328f64b010100050a61737365744f776e657202000001096275726e45706f636801010000096275726e52696768740101000112696e666c6174696f6e416c6c6f77616e6365020100010b7570646174655269676874010100000001467bac8139b1f513fd37b31e74459d860ab2f67033da9ad472e6da3d1ff38a7b04076372656174656400010001000c697373756564537570706c79000100010004737065630001000100057465726d730001000100040a61737365744f776e6572000000ffff096275726e45706f6368000000010012696e666c6174696f6e416c6c6f77616e6365000000ffff0b75706461746552696768740000000100000301030406044275726e0101318ecaa072870bd95230a9515a6aaa14982bac9e3fb15ea2528eb0405227a7a1010c6275726e6564537570706c79000100010001047573656401096275726e526967687401000100010666757475726501096275726e5269676874000001000003010305000549737375650101467bac8139b1f513fd37b31e74459d860ab2f67033da9ad472e6da3d1ff38a7b010c697373756564537570706c7900010001000104757365640112696e666c6174696f6e416c6c6f77616e63650100ffff020b62656e6566696369617279010a61737365744f776e65720000ffff066675747572650112696e666c6174696f6e416c6c6f77616e63650000ffff000401030406010b62656e6566696369617279094f70656e45706f636801000001047573656401096275726e45706f63680100010002096275726e52696768740001000100046e65787401096275726e45706f636800000100000001096275726e52696768740652656e616d65010001036e657701047370656301000100010475736564010b7570646174655269676874010001000106667574757265010b75706461746552696768740000010000000106667574757265075265706c6163650101318ecaa072870bd95230a9515a6aaa14982bac9e3fb15ea2528eb0405227a7a1010e7265706c61636564537570706c79000100010001047573656401096275726e526967687401000100020b62656e6566696369617279010a61737365744f776e65720000ffff0666757475726501096275726e526967687400000100000401020305010b62656e6566696369617279085472616e73666572000000010870726576696f7573010a61737365744f776e65720100ffff010b62656e6566696369617279010a61737365744f776e65720100ffff000102010b62656e656669636961727900a1879e8f56bdd6fd8ca323b94f39e060643aa69a39219194d60c4c8a7c7b01c801085472616e7366657200c70931fdf90384c1f4b4f34b2543e80876e35dbf0f15b567ccec8b340b811cc99abcb19a89c3e955b796ff9044c1078e81ab142c309d8524c774094a1835838f0303000763726561746564d0070473706563d107057465726d7301d0070b62656e65666963696172790001d007085472616e7366657200000000c70931fdf90384c1f4b4f34b2543e80876e35dbf0f15b567ccec8b340b811cc98000000303000100040000000000d00701000c00044449424104444942410002d1070100060004004449424101d007010100030177326df3704bdca90c1ad9b33a2e064e88d0cd631bb71ab15619481c2ca92ce7010000002c46352ea481d989080500000000000000073dbed09d95db3df59f8ad1e443bbd0d76f31d804b7eca91f08ab85b1a94ead0000000000000000000000000000";
+const UDA_CONTRACT: &str = "0100000000050300596e7a3a0ea3659e1030f18b4d65a3b913ee585007fcf75e0bb6cdc169c1c03b0100340899624db7e426929954932edfc8d267a5529702293cd5abc0564e6f225b33faf10100350818cb946f1293cf180e9d78dcc65bc59b472ffffeadfbf58db198cc8328f64b01010036080b8789e08890aa2ea1569dbbb6f547f9a435d30db8961fff85fba67deb00174601003808086754f41789c3b935123433d959aa4c5441190bc298e1c8d1a7ed92be8c56840100013408021870edcbec640794b304356cc55c909e7ffd3de152d84c73f2855e5783bbdd5900d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c05030001000100340801000100350801000100360800000100380800000100013408010001000000013408d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0001340801000100013408010001000056000001196d8529c8f6da444d154e8d9c8cbd9e0f8d89025a3bb09121c63d52b01e870501a5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed707fc7e770beea38fc53848e2890f75b66935f005a2e954bd989f4c099392b0a5081cabbfc3d826c0bfd1e9770a889efacc8b6716ad014a3eec10b65915302290420000000000000000ffffffffffffffff086754f41789c3b935123433d959aa4c5441190bc298e1c8d1a7ed92be8c5684060202696477f0d07dccb3bb52a483de90c1e8528ea04e040c04178253411c821e21378f63046e616d65b6ebccff4b9f793259f04b314fb76612808a97c5285cb735ebe39ed8bf8205940b6c500d94264462fb71c0e87478037695b4646158b9ec519dd6d47e74dc54f3035f05737061636520046578636c210671756f7465732204686173682306646f6c6c6172240770657263656e742509616d70657273616e64260a61706f7374726f7068652708627261636b65744c2808627261636b6574522908617374657269736b2a04706c75732b05636f6d6d612c056d696e75732d03646f742e05736c6173682f047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e653905636f6c6f6e3a0973656d69436f6c6f6e3b046c6573733c05657175616c3d07677265617465723e087175657374696f6e3f02617440014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a0a7371427261636b65744c5b096261636b536c6173685c0a7371427261636b6574525d0563617265745e066c6f646173685f086261636b7469636b60016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7a0963427261636b65744c7b04706970657c0963427261636b6574527d0574696c64657e0b8789e08890aa2ea1569dbbb6f547f9a435d30db8961fff85fba67deb001746060805696e64657865015643e4488cf3f714340179d725e828f2c04bfc41cc3b27f6054fda37731d067469636b6572a1f5ff46847f287b7b3063062ee4a342ed6334cf5e0fb2730d48cf7f5bb0037e046e616d65cf8e262506a3a4b2a6abca385a6502abf8036046a6dd8da63ae61d309b3c752a0764657461696c7345b780258601c526b23b5b4861460a9050e13f35fbbe8305a8001157e40138880770726576696577ae5d4de320882d36a046dd72848e756a08638149d751afac47a0b77d8c2733ac056d656469611b2fd8ca8d5f1671165441ae083116506c4e265a8d7cb9869a409db9701dafaa0b6174746163686d656e7473500a5e82892023b23d60a374fea4fd1f6b3bd66fec3f396fbe8aeac319dc9f8c0872657365727665736e74cc9f1d0ccb54ceff2df324c9d455c1acea72323acdb75cce02b7bd97a3550ed21a3b1985f8f55f1affc7261d551f6bab4bf206cdc448fdcb97b7c44809cd0602047574786fe86a40cd5c25f9bde284ff69a51132a7412b2155bd0e5612c1bfe3849560c5060570726f6f66f2ba2480169e7d3a48fd00302d1fdf80b9e3118f01a25ba6752d7dcf81ef0838140cb31d53f4bd9e48091b9b557d8203262f22c26ca7b1688652825b3f8ab23a0501a5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed715dfa1833ae7a0256426366659a343d3ca0308da6670cfd663f19878ef9f7465087a7b8f49025cafad1370f90a4b0bb78028c417521384c039b409386768ed91140000000000000000ffffffffffffffff181aa1e3ed74bd51a3ccdb6cb2180bd413f2f891c28a765f16a2a8f3ba2dabb1086015c0e70764a9f48ae71dd448caf4035bffebdb62eff70dc8794286f8910f41010000000000000008000000000000001870edcbec640794b304356cc55c909e7ffd3de152d84c73f2855e5783bbdd59050265015643e4488cf3f714340179d725e828f2c04bfc41cc3b27f6054fda37731d19ebe69d137048e1596fc1e7e950af0a64c1765f29e41224ab2d2bf4e7605c7718cb946f1293cf180e9d78dcc65bc59b472ffffeadfbf58db198cc8328f64b010501560d96f7a47924b2c3df040e6463398fd65fd591652c294342bfa5f93915515419ebe69d137048e1596fc1e7e950af0a64c1765f29e41224ab2d2bf4e7605c7705015ca149585de534ee91b3e3a030b7efd4cdb79abea9152f101f3759b4c7210e1f1a13cfad8446fb15031758c2dfa6a09d583183eac1afa057e550d636a4e5021b080b6c500d94264462fb71c0e87478037695b4646158b9ec519dd6d47e74dc54f3010000000000000014000000000000001b2fd8ca8d5f1671165441ae083116506c4e265a8d7cb9869a409db9701dafaa040200046e6f6e65d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0104736f6d652b9fd922298de5f15323ff775a1f79d0d35619fe1298ac9047ead0edf29fb66f1cabbfc3d826c0bfd1e9770a889efacc8b6716ad014a3eec10b659153022904200401deddddb38b837060353fb361533ffa352f30397b4c7abc15bbdbd25da8915900501e15eba16722843cc10954a1b33479dbc343c67693fb7ade855e17f656b9f583e21e33e43c5aa5cf2292753a338c167ed3b639c4dd9cebda98f346917bac943750501a5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed72a5baaac5089ff098d150b482cfed8bcd01a91c0d7b45d32216ed576ab71ebdd071cabbfc3d826c0bfd1e9770a889efacc8b6716ad014a3eec10b659153022904220002b9fd922298de5f15323ff775a1f79d0d35619fe1298ac9047ead0edf29fb66f0501faa9838e0ae786f75e697467292235e328e14488d83fcddd79f3c33c025122762d2519390702d9a3484b79e289e7bbf2a8cad7fb5d2bc0a5f8471dba6fa3c03c031a014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a30a0d30c6bbfe91e2490b34278c7bf7f66ee6444011888dc4132bf304c6fe7680807fc7e770beea38fc53848e2890f75b66935f005a2e954bd989f4c099392b0a50000000000000000ffffffffffffffff332de353e616da0cb6394a23d10773fae94ff131f4fe216b2d0001bff1c7b2fb06040776657273696f6ea87cc6796467e557d744e5c9623ae00ccb95d91a0d2f03df74c075124fbec79e06696e7075747315dfa1833ae7a0256426366659a343d3ca0308da6670cfd663f19878ef9f7465076f7574707574738ac60209af0af69c70a282f79ba38ddf4ab5a2afbca5da03dc69ffab00e395ca086c6f636b54696d6535da1d1539206ebbdddca4c33d8731eaf29ba7da74df3dc882201c4e16b5b91c339840868f9caa65acd0874d1a06eb2ee2614cdcd537f77615c836cb9a613d93033f04646173682d047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7a35da1d1539206ebbdddca4c33d8731eaf29ba7da74df3dc882201c4e16b5b91c0501a5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed73c8add893836f8c04c073d1f454d9525cf96a7d7557296fa663ddbaaefa843940334014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7a3f2b72b7c4af1a630cb6d3ff088baf351093ee465b9e7d472a610082e449d7e4050163aa2314e8b147c8b284dfb39a9e10d19caad5faea848e3cb9849d9167d6344a406e55d3696e06af0b904b93f8a6b2baf00b19b278b2df825db46015395cda8f0501a4c8bc8c13e5ed1abd6f635282aa70bbf77f4e3be571a6f9c7882c8a981ec72f45b780258601c526b23b5b4861460a9050e13f35fbbe8305a8001157e4013888040200046e6f6e65d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0104736f6d65caff8faeb38a00a04e3621538f8e61d75a85a465cb0a0e48c3593e7eaa6c5fc7467bac8139b1f513fd37b31e74459d860ab2f67033da9ad472e6da3d1ff38a7b060108726573657276657347b3532053f83fc1ad3abe88453620cb647d86ce5ee4e3de56b5c938423eb9cd47b3532053f83fc1ad3abe88453620cb647d86ce5ee4e3de56b5c938423eb9cd090ed21a3b1985f8f55f1affc7261d551f6bab4bf206cdc448fdcb97b7c44809cd0000000000000000ffff000000000000500a5e82892023b23d60a374fea4fd1f6b3bd66fec3f396fbe8aeac319dc9f8c0a0001faa9838e0ae786f75e697467292235e328e14488d83fcddd79f3c33c02512276000000000000000014000000000000005014105f0c4026ca12abae377de26f8e3ca0608dd36c483ed70f47195cd334d4050184de03c07d1d477bb76d40a947d26e1ed87dc44eaf86720dd4dc8a48047373b3560d96f7a47924b2c3df040e6463398fd65fd591652c294342bfa5f93915515408fba958721a3d335406b368c36f5a82790960cce239febcafe189ba9839d5da780000000000000000ffff000000000000596e7a3a0ea3659e1030f18b4d65a3b913ee585007fcf75e0bb6cdc169c1c03b05016632da64bbf5bc3cc9f522452f5377a83804cdcd95006bc467a1394cc1a2b48f5be69aabb827ae7a16a0c65f72ee867e327e63039aa8d51b176a5a1410b2e269050107fc7e770beea38fc53848e2890f75b66935f005a2e954bd989f4c099392b0a55ca149585de534ee91b3e3a030b7efd4cdb79abea9152f101f3759b4c7210e1f00085d03c4178da98e7e3f3af343e3997d74201d11f42732cfbea2b04b8e3ff15f2203130b696e646976697369626c65000464656369010563656e746902056d696c6c690309646563694d696c6c69040a63656e74694d696c6c6905056d6963726f0609646563694d6963726f070a63656e74694d6963726f08046e616e6f0908646563694e616e6f0a0963656e74694e616e6f0b047069636f0c08646563695069636f0d0963656e74695069636f0e0566656d746f0f096465636946656d746f100a63656e746946656d746f11046174746f125f7ae1f33fc70d8ba4a8945811223ef9b35135ada0240f39b0dc282ee9e48b8f05015014105f0c4026ca12abae377de26f8e3ca0608dd36c483ed70f47195cd334d46015c0e70764a9f48ae71dd448caf4035bffebdb62eff70dc8794286f8910f410324047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a618622d17baef06602dfe775c980e0b36b6ebae8090d8075f7513bec4b93f64d03020566616c73650004747275650163aa2314e8b147c8b284dfb39a9e10d19caad5faea848e3cb9849d9167d6344a08fba958721a3d335406b368c36f5a82790960cce239febcafe189ba9839d5da780100000000000000ff0000000000000065015643e4488cf3f714340179d725e828f2c04bfc41cc3b27f6054fda37731d0501a5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed765eb913c17e7e03d9bb1c615293b085ade1511a7ba1c6b30160a33722231a10a06020576616c756597f5d70642ae2a3392253ba84e1dcec493e3bf3ed371b1877386354ecc88826b0c7363726970745075626b6579cae95645ac9ae8fb1892420d1c8b46c220aac64ce6dc8258866fa365aea490436632da64bbf5bc3cc9f522452f5377a83804cdcd95006bc467a1394cc1a2b48f00446b87cc20476a4daa93590b2568db9e0ed54ab8c43c8e938bb72f4e1ad7f1e9de0308106672616374696f6e4f766572666c6f77010e6e6f6e457175616c56616c756573020c696e76616c696450726f6f660314696e73756666696369656e7452657365727665730415697373756545786365656473416c6c6f77616e636506126e6f6e4672616374696f6e616c546f6b656e07126e6f6e456e6772617661626c65546f6b656e0815696e76616c69644174746163686d656e7454797065096e74cc9f1d0ccb54ceff2df324c9d455c1acea72323acdb75cce02b7bd97a355040200046e6f6e65d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0104736f6d657d654e97ed96167c3c3c491357425127b6c0f008a2cd70541773fa1b2d4daa24728ea68a97de7420f62a5969b04b872604bf45d6df38959c7e26e7a72490e4ad033f047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a066c6f646173685f016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7a77f0d07dccb3bb52a483de90c1e8528ea04e040c04178253411c821e21378f6300017a7b8f49025cafad1370f90a4b0bb78028c417521384c039b409386768ed911406040a707265764f7574707574e86a40cd5c25f9bde284ff69a51132a7412b2155bd0e5612c1bfe3849560c50609736967536372697074cbb261c09079bea5f78800d747feb5f8c024e2c04f73f945b468068df143ec940873657175656e636501196d8529c8f6da444d154e8d9c8cbd9e0f8d89025a3bb09121c63d52b01e87077769746e657373ef9b75968c7d8d6905f17c692061d6e93a2abe8248e6a70b0867a2dffe84fb257d654e97ed96167c3c3c491357425127b6c0f008a2cd70541773fa1b2d4daa2405010ed21a3b1985f8f55f1affc7261d551f6bab4bf206cdc448fdcb97b7c44809cd8057ae07f47b923accc5361fcf40d92982e841f757dd59b16d7bdbc8eabe99a50501181aa1e3ed74bd51a3ccdb6cb2180bd413f2f891c28a765f16a2a8f3ba2dabb184de03c07d1d477bb76d40a947d26e1ed87dc44eaf86720dd4dc8a48047373b3080b6c500d94264462fb71c0e87478037695b4646158b9ec519dd6d47e74dc54f3010000000000000028000000000000008ac60209af0af69c70a282f79ba38ddf4ab5a2afbca5da03dc69ffab00e395ca0865eb913c17e7e03d9bb1c615293b085ade1511a7ba1c6b30160a33722231a10a0000000000000000ffffffffffffffff97f5d70642ae2a3392253ba84e1dcec493e3bf3ed371b1877386354ecc88826b05015ca149585de534ee91b3e3a030b7efd4cdb79abea9152f101f3759b4c7210e1f99624db7e426929954932edfc8d267a5529702293cd5abc0564e6f225b33faf10602066e616d696e67de7c753c6bd56774f87245306ca463409710cee01bedca9120fcaeef93bff58609707265636973696f6e5d03c4178da98e7e3f3af343e3997d74201d11f42732cfbea2b04b8e3ff15f229e14cce6b185cca477ba44634bfa2c5871a97e0c32b7aec1df6a94522a9319c70310025f3000025f3101025f3202025f3303025f3404025f3505025f3606025f3707025f3808025f3909035f31300a035f31310b035f31320c035f31330d035f31340e035f31350fa1f5ff46847f287b7b3063062ee4a342ed6334cf5e0fb2730d48cf7f5bb0037e040200046e6f6e65d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0104736f6d65e93845aac316b6d6d0642da28663d3f803a0767a405d73297b36be8bd911a46da38242f3cbd7d1d0a112da8de1602f56d4f86f0ec4c476c5c068594c4b04655605012a5baaac5089ff098d150b482cfed8bcd01a91c0d7b45d32216ed576ab71ebdda4c8bc8c13e5ed1abd6f635282aa70bbf77f4e3be571a6f9c7882c8a981ec72f06020474797065fb6b2ec4700e9bf1cd0f24bb9ced0dd78dbb4e1dcaf5c68fa6935a7a4fdc69420464617461f2ba2480169e7d3a48fd00302d1fdf80b9e3118f01a25ba6752d7dcf81ef0838a5cffb327513077e4f1fc96072928244a01ea18c6cdaf598da0b5735ac075ed70004a62e384a135184183b9da6b9796b06e88c75c6803139dd391148ea049ae3d7e10310047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e65390374656e6106656c6576656e62067477656c76656308746869727465656e6408666f75727465656e65076669667465656e66a7ccf154860d368a2b95e87ad394d13d7e6430ac493c7bcdbb95a1c5aea1ccb00602096170706c696564546f65015643e4488cf3f714340179d725e828f2c04bfc41cc3b27f6054fda37731d07636f6e74656e74a4c8bc8c13e5ed1abd6f635282aa70bbf77f4e3be571a6f9c7882c8a981ec72fa87cc6796467e557d744e5c9623ae00ccb95d91a0d2f03df74c075124fbec79e05016632da64bbf5bc3cc9f522452f5377a83804cdcd95006bc467a1394cc1a2b48fae5d4de320882d36a046dd72848e756a08638149d751afac47a0b77d8c2733ac040200046e6f6e65d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0104736f6d65406e55d3696e06af0b904b93f8a6b2baf00b19b278b2df825db46015395cda8fb6ebccff4b9f793259f04b314fb76612808a97c5285cb735ebe39ed8bf82059405011a13cfad8446fb15031758c2dfa6a09d583183eac1afa057e550d636a4e5021bbaa0e3fcab9d55d168bb1e26837153afe89f83a27e0498554d3963f6431fd3f4033e047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539014141014242014343014444014545014646014747014848014949014a4a014b4b014c4c014d4d014e4e014f4f015050015151015252015353015454015555015656015757015858015959015a5a016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7abc2891b1c66ac5f5e61059a32077daa82333a1435bd34608b30b03fc017d9545030a047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e6539c0ec009baa44e89cdbeb95c6f18117a53618352f070bc59fc1d8e49f07b4298c040200046e6f6e65d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0104736f6d651deddddb38b837060353fb361533ffa352f30397b4c7abc15bbdbd25da891590c3cad9212706689af885a1b06a8d27ed3d3ff250c409f93fee32eb32544866340360035f333220035f333321035f333422035f333523035f333624035f333725035f333826035f333927035f343028035f343129035f34322a035f34332b035f34342c035f34352d035f34362e035f34372f035f343830035f343931035f353032035f353133035f353234035f353335035f353436035f353537035f353638035f353739035f35383a035f35393b035f36303c035f36313d035f36323e035f36333f035f363440035f363541035f363642035f363743035f363844035f363945035f373046035f373147035f373248035f373349035f37344a035f37354b035f37364c035f37374d035f37384e035f37394f035f383050035f383151035f383252035f383353035f383454035f383555035f383656035f383757035f383858035f383959035f39305a035f39315b035f39325c035f39335d035f39345e035f39355f035f393660035f393761035f393862035f393963045f31303064045f31303165045f31303266045f31303367045f31303468045f31303569045f3130366a045f3130376b045f3130386c045f3130396d045f3131306e045f3131316f045f31313270045f31313371045f31313472045f31313573045f31313674045f31313775045f31313876045f31313977045f31323078045f31323179045f3132327a045f3132337b045f3132347c045f3132357d045f3132367e045f3132377fcae95645ac9ae8fb1892420d1c8b46c220aac64ce6dc8258866fa365aea4904305015be69aabb827ae7a16a0c65f72ee867e327e63039aa8d51b176a5a1410b2e269caff8faeb38a00a04e3621538f8e61d75a85a465cb0a0e48c3593e7eaa6c5fc705013f2b72b7c4af1a630cb6d3ff088baf351093ee465b9e7d472a610082e449d7e4cbb261c09079bea5f78800d747feb5f8c024e2c04f73f945b468068df143ec9405015be69aabb827ae7a16a0c65f72ee867e327e63039aa8d51b176a5a1410b2e269cf8e262506a3a4b2a6abca385a6502abf8036046a6dd8da63ae61d309b3c752a040200046e6f6e65d83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0104736f6d655f7ae1f33fc70d8ba4a8945811223ef9b35135ada0240f39b0dc282ee9e48b8fd83fbee02f0de5b46cf80fe11ef7fdf061c78d975d31ade9eea2bc4099339e6c0000da57a3e7de97d8c4dcaafbd143a882a7587e35c3fa805625098c97e71629a5b408c3cad9212706689af885a1b06a8d27ed3d3ff250c409f93fee32eb325448663401000000000000004000000000000000de7c753c6bd56774f87245306ca463409710cee01bedca9120fcaeef93bff5860603067469636b65728057ae07f47b923accc5361fcf40d92982e841f757dd59b16d7bdbc8eabe99a5046e616d655014105f0c4026ca12abae377de26f8e3ca0608dd36c483ed70f47195cd334d40764657461696c7345b780258601c526b23b5b4861460a9050e13f35fbbe8305a8001157e4013888e15eba16722843cc10954a1b33479dbc343c67693fb7ade855e17f656b9f583e0501da57a3e7de97d8c4dcaafbd143a882a7587e35c3fa805625098c97e71629a5b4e86a40cd5c25f9bde284ff69a51132a7412b2155bd0e5612c1bfe3849560c50606020474786964a38242f3cbd7d1d0a112da8de1602f56d4f86f0ec4c476c5c068594c4b04655604766f757421e33e43c5aa5cf2292753a338c167ed3b639c4dd9cebda98f346917bac94375e93845aac316b6d6d0642da28663d3f803a0767a405d73297b36be8bd911a46d05018057ae07f47b923accc5361fcf40d92982e841f757dd59b16d7bdbc8eabe99a5eedec3ebaa08fb010f2f96e17a0e039ad87fe60771a7301f8c5e88995d05676b0310047a65726f30036f6e65310374776f320574687265653304666f757234046669766535037369783605736576656e3705656967687438046e696e65390374656e4106656c6576656e42067477656c76654308746869727465656e4408666f75727465656e45076669667465656e46ef9b75968c7d8d6905f17c692061d6e93a2abe8248e6a70b0867a2dffe84fb25050130a0d30c6bbfe91e2490b34278c7bf7f66ee6444011888dc4132bf304c6fe768f2ba2480169e7d3a48fd00302d1fdf80b9e3118f01a25ba6752d7dcf81ef0838081cabbfc3d826c0bfd1e9770a889efacc8b6716ad014a3eec10b65915302290420000000000000000ffff000000000000f9170804ddae0479f8d5af74ab3bd202e6ea4172d9a9b93707151adb7fc40ca1031a016161016262016363016464016565016666016767016868016969016a6a016b6b016c6c016d6d016e6e016f6f017070017171017272017373017474017575017676017777017878017979017a7afaa9838e0ae786f75e697467292235e328e14488d83fcddd79f3c33c0251227606020474797065086754f41789c3b935123433d959aa4c5441190bc298e1c8d1a7ed92be8c5684066469676573742a5baaac5089ff098d150b482cfed8bcd01a91c0d7b45d32216ed576ab71ebddfb6b2ec4700e9bf1cd0f24bb9ced0dd78dbb4e1dcaf5c68fa6935a7a4fdc694206030474797065e15eba16722843cc10954a1b33479dbc343c67693fb7ade855e17f656b9f583e0773756274797065c0ec009baa44e89cdbeb95c6f18117a53618352f070bc59fc1d8e49f07b4298c0763686172736574c0ec009baa44e89cdbeb95c6f18117a53618352f070bc59fc1d8e49f07b4298cfba958721a3d335406b368c36f5a82790960cce239febcafe189ba9839d5da78010000000001e6b0c7ab8033b92ebfff89bda3e3a4403a7c37454e4d1de2d78f7af8fe83008600055247423231060f6174746163686d656e74547970657301086754f41789c3b935123433d959aa4c5441190bc298e1c8d1a7ed92be8c56840001076372656174656401596e7a3a0ea3659e1030f18b4d65a3b913ee585007fcf75e0bb6cdc169c1c03b01000a656e67726176696e677301a7ccf154860d368a2b95e87ad394d13d7e6430ac493c7bcdbb95a1c5aea1ccb0000104737065630199624db7e426929954932edfc8d267a5529702293cd5abc0564e6f225b33faf10100057465726d730118cb946f1293cf180e9d78dcc65bc59b472ffffeadfbf58db198cc8328f64b01010006746f6b656e73010b8789e08890aa2ea1569dbbb6f547f9a435d30db8961fff85fba67deb0017460001030a61737365744f776e6572051870edcbec640794b304356cc55c909e7ffd3de152d84c73f2855e5783bbdd5900000112696e666c6174696f6e416c6c6f77616e636505140cb31d53f4bd9e48091b9b557d8203262f22c26ca7b1688652825b3f8ab23a0100010b7570646174655269676874010100000001467bac8139b1f513fd37b31e74459d860ab2f67033da9ad472e6da3d1ff38a7b050f6174746163686d656e745479706573000000ffff0763726561746564000100010004737065630001000100057465726d73000100010006746f6b656e73000000ffff030a61737365744f776e6572000000ffff12696e666c6174696f6e416c6c6f77616e6365000000ffff0b757064617465526967687400000001000004010304090407456e67726176650100010a656e67726176696e67730001000100010870726576696f7573010a61737365744f776e65720100ffff010b62656e6566696369617279010a61737365744f776e65720100ffff000401020708010b62656e65666963696172790549737375650101467bac8139b1f513fd37b31e74459d860ab2f67033da9ad472e6da3d1ff38a7b02126e65774174746163686d656e745479706573010f6174746163686d656e7454797065730000ffff096e6577546f6b656e730106746f6b656e730000ffff0104757365640112696e666c6174696f6e416c6c6f77616e63650100ffff020b62656e6566696369617279010a61737365744f776e65720000ffff066675747572650112696e666c6174696f6e416c6c6f77616e63650000ffff00050103040609010b62656e65666963696172790652656e616d65010001036e657701047370656301000100010475736564010b7570646174655269676874010001000106667574757265010b75706461746552696768740000010000000106667574757265085472616e73666572000000010870726576696f7573010a61737365744f776e65720100ffff010b62656e6566696369617279010a61737365744f776e65720100ffff0003010207010b62656e6566696369617279006b87cc20476a4daa93590b2568db9e0ed54ab8c43c8e938bb72f4e1ad7f1e9de01085472616e7366657200ca0c967f8bb57d6d85c8a0d0e1ac58816aa9a8d503bb279c50d8e37655f94c73e6b0c7ab8033b92ebfff89bda3e3a4403a7c37454e4d1de2d78f7af8fe8300860403000763726561746564340804737065633508057465726d73360806746f6b656e730134080b62656e656669636961727900013408085472616e7366657200000000ca0c967f8bb57d6d85c8a0d0e1ac58816aa9a8d503bb279c50d8e37655f94c738000000403000100040000000000340801000c0004444942410444494241000235080100060004004449424136080100400001000000010444494241010444494241000105696d6167650103706e67001d0068747470733a2f2f636172626f6e61646f2e696f2f646962612e706e6700000001340802010003019ee7b29502b09271f6e1f7222fdb5c276d4df6c2b4cd4e33ff960e21e353ae9401000000b4d3d37e2fe0bbb40c000100000001000000000000000000000000000000000000000000";
